requirements.txt
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
cmake==4.0.2
contourpy==1.3.2
cycler==0.12.1
filelock==3.18.0
Flask==3.1.1
fonttools==4.58.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.8
lit==18.1.8
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
networkx==3.4.2
numpy==1.24.3
packaging==25.0
pillow==11.2.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
requests==2.32.3
six==1.17.0
sympy==1.14.0
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
triton==2.0.0
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.1.3
